+++ src/reference/make_reference/script.sh
@@ -20,13 +20,21 @@ trap clean_up EXIT
 
 echo "> Processing genome sequence"
 genome_fasta="$tmpdir/genome_sequence.fa"
-# curl "$par_genome_fasta" | gunzip > "$genome_fasta"
-gunzip -c "$par_genome_fasta" > "$genome_fasta"
+# if genome is gzipped, extract. otherwise not
+if [[ "$par_genome_fasta" == *.gz ]]; then
+  gunzip -c "$par_genome_fasta" > "$genome_fasta"
+else
+  cp "$par_genome_fasta" "$genome_fasta"
+fi
 
 echo "> Processing transcriptome annotation"
 transcriptome_gtf="$tmpdir/transcriptome_annotation.gtf"
-# curl "$par_transcriptome_gtf" | gunzip > "$transcriptome_gtf"
-gunzip -c "$par_transcriptome_gtf"> "$transcriptome_gtf"
+# if transcriptome is gzipped, extract. otherwise not
+if [[ "$par_transcriptome_gtf" == *.gz ]]; then
+  gunzip -c "$par_transcriptome_gtf" > "$transcriptome_gtf"
+else
+  cp "$par_transcriptome_gtf" "$transcriptome_gtf"
+fi
 
 if [[ ! -z $par_ercc ]]; then
   echo "> Processing ERCC sequences"